requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
requests
sendgrid
